Move `InlineContentBuilder` and `whitespace::*` into separate file
